Factor the Task & User classes out of index.js.
