+++ src/index.js
@@ -4,7 +4,9 @@ const fs = require('fs');
 const http = require('http');
 const url = require('url');
 const querystring = require('querystring');
-const crypto = require('crypto');
+
+const Task = require('./task.js');
+const User = require('./user.js');
 
 const hostname = '127.0.0.1';
 const port = 3000;
@@ -88,16 +90,8 @@ function backupTasks(request, response) {
         const tasks = JSON.parse(body);
         const bytes = Buffer.byteLength(body);
         console.log(tasks, bytes);
-        setTasks(tasks);
+        Task.setTasks(tasks);
         response.end(`Received ${bytes} bytes.`);
-    });
-}
-
-function setTasks(tasks) {
-    fs.writeFile('../data/tasks.json', JSON.stringify(tasks), error => {
-        if (error) {
-            console.error(error);
-        }
     });
 }
 
@@ -105,54 +99,6 @@ function createAccount(request, response) {
     handlePostRequest(request, response, body => {
         const data = querystring.parse(body);
         const user = { name: data.name }
-        createUser(user, data.password);
+        User.create(user, data.password);
     });
-}
-
-function getUsers() {
-    try {
-        const json = fs.readFileSync('../data/users.json', 'utf8');
-        return JSON.parse(json);
-    }
-    catch (e) {
-        console.log(e);
-    }
-
-    return {};
-}
-
-function createUser(user, password) {
-    const users = getUsers();
-
-    if (user.name in users) {
-        console.log(`The username ${user.name} already exists.`);
-        return;
-    }
-
-    user.salt = createSalt();
-    user.hash = hashPassword(password, user.salt);
-    user.created = Date.now();
-    console.log(user);
-
-    users[user.name] = user;
-    setUsers(users);
-}
-
-function setUsers(users) {
-    fs.writeFile('../data/users.json', JSON.stringify(users), error => {
-        if (error) {
-            console.error(error);
-        }
-    });
-}
-
-function createSalt() {
-    return crypto.randomBytes(16).toString('hex');
-}
-
-function hashPassword(password, salt) {
-    const iterations = 1000;
-    const keylen = 64;
-    password = password.toString().normalize();
-    return crypto.pbkdf2Sync(password, salt, iterations, keylen, 'sha512').toString('hex');
 }
